VideoDeduplicator/requirements.txt
opencv-python-headless
imagehash
Pillow
fastapi
uvicorn[standard]
jinja2
python-multipart